After some changes requested! 1 out of 4 changes done!

+++ src/main/java/firsttask/taskmanager/Controller/UserController.java
@@ -1,42 +1,31 @@
 package firsttask.taskmanager.Controller;
 
+import firsttask.taskmanager.Exceptions.GeneralException;
 import firsttask.taskmanager.Exceptions.UserNotFoundException;
 import firsttask.taskmanager.Repositories.TaskRepository;
 import firsttask.taskmanager.Repositories.UserRepository;
-import firsttask.taskmanager.assemblers.TaskModelAssembler;
-import firsttask.taskmanager.assemblers.UserModelAssembler;
 import firsttask.taskmanager.domain.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.hateoas.CollectionModel;
-import org.springframework.hateoas.EntityModel;
-import org.springframework.hateoas.IanaLinkRelations;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
-import java.util.stream.Collectors;
 
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
-import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;
 
+@RestController
 public class UserController {
 
 
     private final TaskRepository taskRepository;
     private final UserRepository userRepository;
-    private final UserModelAssembler userModelAssembler;
-    private final TaskModelAssembler taskModelAssembler;
-    private static Logger logger = LoggerFactory.getLogger(UserController.class);
+    private static final Logger logger = LoggerFactory.getLogger(UserController.class);
 
 
-    public UserController(TaskRepository taskRepository, UserRepository userRepository,UserModelAssembler UserAssembler, TaskModelAssembler TaskAssembler) {
+    public UserController(TaskRepository taskRepository, UserRepository userRepository) throws GeneralException {
         this.taskRepository = taskRepository;
         this.userRepository = userRepository;
-        this.userModelAssembler=UserAssembler;
-        this.taskModelAssembler=TaskAssembler;
         logger.trace("creating Repositories and assemblers ");
     }
 
@@ -44,23 +33,20 @@ public class UserController {
 
     //return all the users in that database
     @GetMapping("/users")
-    public CollectionModel<EntityModel<User>> returnAllUsers() {
+    public  List<User> returnAllUsers() throws GeneralException {
         logger.info("A get all users request initialized ");
-        System.out.println("inside get all");
-        List<EntityModel<User>> users = userRepository.findAll().stream() //
-                .map(userModelAssembler::toModel) //
-                .collect(Collectors.toList());
+
+        List<User> users = userRepository.findAll();
         logger.trace("retrieving all the users " );
-        return CollectionModel.of(users, linkTo(methodOn(UserController.class).returnAllUsers()).withSelfRel());
+        return users;
     }
 
 
 
     // return a  user and his task
     @GetMapping("/users/{id}")
-    public /*EntityModel<User>*/ User returnUser(@PathVariable Long id)  {
+    public  User returnUser(@PathVariable Long id)  throws GeneralException {
         logger.info("A get userwith id "+ id  +" request initialized ");
-        System.out.println("inside the get method ");
         User user = userRepository.findById(id) //
                 .orElseThrow(() -> new UserNotFoundException(id));
         logger.trace("retrieving the user with id :  "+ id  );
@@ -71,52 +57,42 @@ public class UserController {
 
     //creating user
     @PostMapping("/users")
-    ResponseEntity<EntityModel<User>> createNewUser(@RequestBody User newUser) {
+    User createNewUser(@RequestBody User newUser) throws GeneralException {
         logger.info("A create user request initialized ");
-        System.out.println("inside the post");
-        EntityModel<User> userEntityModel = userModelAssembler.toModel(userRepository.save(newUser));
-        logger.trace("Creating new user "+ userEntityModel);
-        return ResponseEntity //
-                .created(userEntityModel.getRequiredLink(IanaLinkRelations.SELF).toUri()) //
-                .body(userEntityModel);
+        User createdUser = userRepository.save(newUser);
+        logger.trace("Creating new user "+ createdUser);
+        return createdUser;
     }
 
 
 
     //Editing existing user after making sure of his password
     @PutMapping("/users/{id}")
-    ResponseEntity<?> edUser(@RequestBody User editUser, @PathVariable Long id) {
+    User edUser(@RequestBody User editUser, @PathVariable Long id) throws GeneralException {
         logger.info("A update user request initialized ");
         User updatedUser = userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
-        System.out.println("inisde post");
-            updatedUser.setId(editUser.getId());
-            updatedUser.setName(editUser.getName());
-            updatedUser.setEmail(editUser.getEmail());
-            updatedUser.setAge(editUser.getAge());
-            EntityModel<User> userEntityModel = userModelAssembler.toModel(updatedUser);
-            logger.trace("updating user information "+userEntityModel);
-            return ResponseEntity //
-                    .created(userEntityModel.getRequiredLink(IanaLinkRelations.SELF).toUri()) //
-                    .body(userEntityModel);
+        updatedUser.setId(editUser.getId());
+        updatedUser.setName(editUser.getName());
+        updatedUser.setEmail(editUser.getEmail());
+        updatedUser.setAge(editUser.getAge());
 
+        logger.trace("updating user information " + updatedUser);
+        return updatedUser;
     }
 
 
 
     //Delete the user by his id and verify the password of that user
     @DeleteMapping("/users/{id}")
-    void deleteUser(@PathVariable Long id,HttpServletResponse response) throws IOException {
-        logger.info("A Delete user request initialized ");
-        User user = userRepository.findById(id) //
-                .orElseThrow(() -> new UserNotFoundException(id));
-        System.out.println("inside delete ");
-            userRepository.deleteById(id);
-            taskRepository.deleteAllByUser_Id(id);
-            logger.trace("Redirecting to the /Users page after deleting a user with id : "+ id );
-            response.sendRedirect("");
+    void deleteUser(@PathVariable Long id,HttpServletResponse response) throws IOException, GeneralException {
 
-
+          logger.info("A Delete user request initialized ");
+          if (userRepository.existsById(id)) {
+              userRepository.deleteById(id);
+              taskRepository.deleteAllByUser_Id(id);
+          }
+          logger.trace("Redirecting to the /Users page after deleting a user with id : " + id);
+          response.sendRedirect("");
 
     }
-
 }
